+++ Reika/SatisfactoryPlanner/GUI/GuiUtil.java
@@ -453,60 +453,60 @@ public class GuiUtil {
 
 	}
 
-	public static String formatProductionDecimal(float amt) {
+	public static String formatProductionDecimal(double amt) {
 		DecimalFormat df = new DecimalFormat("0", DecimalFormatSymbols.getInstance(Locale.ENGLISH));
 		df.setMaximumFractionDigits(4);
 		return df.format(amt);
 	}
 
-	public static final GuiInstance<ItemRateController> createItemView(Consumable c, float baseAmount, ExpandingTilePane container) {
+	public static final GuiInstance<ItemRateController> createItemView(Consumable c, double baseAmount, ExpandingTilePane container) {
 		return createItemView(c, baseAmount, true, inner -> container.addEntry(inner));
 	}
 
-	public static final GuiInstance<ItemRateController> createItemView(Consumable c, float baseAmount, Pane container) {
+	public static final GuiInstance<ItemRateController> createItemView(Consumable c, double baseAmount, Pane container) {
 		return createItemView(c, baseAmount, false, inner -> container.getChildren().add(inner.rootNode));
 	}
 
-	public static final GuiInstance<ItemRateController> createItemView(Consumable c, float baseAmount, TabPane container) {
+	public static final GuiInstance<ItemRateController> createItemView(Consumable c, double baseAmount, TabPane container) {
 		return createItemView(c, baseAmount, false, inner -> {Tab t = new Tab(); t.setContent(inner.rootNode); container.getTabs().add(t);});
 	}
 
-	public static final GuiInstance<ItemRateController> createItemView(Consumable c, float baseAmount, Tab container) {
+	public static final GuiInstance<ItemRateController> createItemView(Consumable c, double baseAmount, Tab container) {
 		return createItemView(c, baseAmount, false, inner -> {container.setContent(inner.rootNode);});
 	}
 
-	public static final GuiInstance<ItemRateController> createItemView(Consumable c, float baseAmount, GridPane container, int col, int row) {
+	public static final GuiInstance<ItemRateController> createItemView(Consumable c, double baseAmount, GridPane container, int col, int row) {
 		return createItemView(c, baseAmount, false, inner -> container.add(inner.rootNode, col, row));
 	}
 
-	public static GuiInstance<ItemRateController> createItemView(Consumable c, float baseAmount, boolean table, Consumer<GuiInstance<ItemRateController>> acceptor) {
+	public static GuiInstance<ItemRateController> createItemView(Consumable c, double baseAmount, boolean table, Consumer<GuiInstance<ItemRateController>> acceptor) {
 		ItemRateController view = new ItemRateController(c, baseAmount, table);
 		GuiInstance<ItemRateController> gui = new GuiInstance<ItemRateController>(view.getRootNode(), view);
 		acceptor.accept(gui);
 		return gui;
 	}
 
-	public static final GuiInstance<ItemCountController> addIconCount(Resource c, float amt, int digits, boolean compress, ExpandingTilePane container) {
+	public static final GuiInstance<ItemCountController> addIconCount(Resource c, double amt, int digits, boolean compress, ExpandingTilePane container) {
 		return addIconCount(c, amt, digits, compress, inner -> container.addEntry(inner));
 	}
 
-	public static final GuiInstance<ItemCountController> addIconCount(Resource c, float amt, int digits, boolean compress, Pane container) {
+	public static final GuiInstance<ItemCountController> addIconCount(Resource c, double amt, int digits, boolean compress, Pane container) {
 		return addIconCount(c, amt, digits, compress, inner -> container.getChildren().add(inner.rootNode));
 	}
 
-	public static final GuiInstance<ItemCountController> addIconCount(Resource c, float amt, int digits, boolean compress, TabPane container) {
+	public static final GuiInstance<ItemCountController> addIconCount(Resource c, double amt, int digits, boolean compress, TabPane container) {
 		return addIconCount(c, amt, digits, compress, inner -> {Tab t = new Tab(); t.setContent(inner.rootNode); container.getTabs().add(t);});
 	}
 
-	public static final GuiInstance<ItemCountController> addIconCount(Resource c, float amt, int digits, boolean compress, Tab container) {
+	public static final GuiInstance<ItemCountController> addIconCount(Resource c, double amt, int digits, boolean compress, Tab container) {
 		return addIconCount(c, amt, digits, compress, inner -> {container.setContent(inner.rootNode);});
 	}
 
-	public static final GuiInstance<ItemCountController> addIconCount(Resource c, float amt, int digits, boolean compress, GridPane container, int col, int row) {
+	public static final GuiInstance<ItemCountController> addIconCount(Resource c, double amt, int digits, boolean compress, GridPane container, int col, int row) {
 		return addIconCount(c, amt, digits, compress, inner -> container.add(inner.rootNode, col, row));
 	}
 
-	public static GuiInstance<ItemCountController> addIconCount(Resource r, float amt, int digits, boolean compress, Consumer<GuiInstance<ItemCountController>> acceptor) {
+	public static GuiInstance<ItemCountController> addIconCount(Resource r, double amt, int digits, boolean compress, Consumer<GuiInstance<ItemCountController>> acceptor) {
 		ItemCountController c = new ItemCountController(r, amt, compress);
 		GuiInstance<ItemCountController> gui = new GuiInstance<ItemCountController>(c.getRootNode(), c);
 		gui.controller.setMaxLength(digits);
